handling chain file
